backend/requirements.txt
fastapi
uvicorn
twilio
python-dotenv
google-generativeai
pydantic
sqlalchemy
aiosqlite
httpx
jinja2
python-multipart
kokoro>=0.4.0
soundfile
numpy
